% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eNN.R
\name{eNN}
\alias{eNN}
\title{Network Reconstruction via epsilon-NN}
\usage{
eNN(E_g)
}
\arguments{
\item{E_g}{Embedding representations of genes.}
}
\value{
The epsilon-NN network.
}
\description{
Reconstruction of gene-gene network via low-dimentional projections (via epsilon-NN).
}
